--- src/main/java/ba/unsa/etf/si/payment/response/Post.java
@@ -1,54 +1,0 @@
-package ba.unsa.etf.si.payment.response;
-
-import java.io.Serializable;
-
-public class Post implements Serializable {
-
-    private int userId;
-    private int id;
-    private String title;
-    private String body;
-
-    public Post(int userId, int id, String title, String body) {
-        this.userId = userId;
-        this.id = id;
-        this.title = title;
-        this.body = body;
-    }
-
-    public Post() {
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getBody() {
-        return body;
-    }
-
-    public void setBody(String body) {
-        this.body = body;
-    }
-
-    public int getUserId() {
-        return userId;
-    }
-
-    public void setUserId(int userId) {
-        this.userId = userId;
-    }
-    // getters and setters
-}
