+++ src/main/java/ba/unsa/etf/si/payment/response/transactionResponse/TransactionSubmitResponse.java
@@ -1,4 +1,4 @@
-package ba.unsa.etf.si.payment.response;
+package ba.unsa.etf.si.payment.response.transactionResponse;
 
 import java.util.UUID;
